Test each client component using runClientTestsForComponent() test runner
